+++ src/holmes/api/__init__.py
@@ -1,0 +1,9 @@
+from . import api, calibration, projection, simulation, utils
+
+__all__ = [
+    "api",
+    "calibration",
+    "projection",
+    "simulation",
+    "utils",
+]